+++ src/quacc/calculators/vasp/vasp.py
@@ -58,7 +58,7 @@ class Vasp(Vasp_):
         input_atoms: Atoms,
         preset: str | Path | None = None,
         use_custodian: bool | DefaultSetting = QuaccDefault,
-        incar_copilot: Literal["off", "critical", "standard", "aggressive"]
+        incar_copilot_mode: Literal["off", "critical", "standard", "aggressive"]
         | DefaultSetting = QuaccDefault,
         copy_magmoms: bool | DefaultSetting = QuaccDefault,
         preset_mag_default: float | DefaultSetting = QuaccDefault,
@@ -70,7 +70,7 @@ class Vasp(Vasp_):
             | None
         ) = None,
         auto_dipole: bool | None = None,
-        **kwargs,
+        **kwargs: Any,
     ) -> None:
         """
         Initialize the VASP calculator.
@@ -88,7 +88,7 @@ class Vasp(Vasp_):
             override any corresponding preset values.
         use_custodian
             Whether to use Custodian to run VASP. Default is True in settings.
-        incar_copilot
+        incar_copilot_mode
             Controls VASP co-pilot mode for automated INCAR parameter handling.
 
             Options include:
@@ -135,10 +135,10 @@ class Vasp(Vasp_):
             if use_custodian == QuaccDefault
             else use_custodian
         )
-        incar_copilot = (
-            self._settings.VASP_INCAR_COPILOT
-            if incar_copilot == QuaccDefault
-            else incar_copilot
+        incar_copilot_mode = (
+            self._settings.VASP_INCAR_COPILOT_MODE
+            if incar_copilot_mode == QuaccDefault
+            else incar_copilot_mode
         )
         copy_magmoms = (
             self._settings.VASP_COPY_MAGMOMS
@@ -158,7 +158,7 @@ class Vasp(Vasp_):
         self.input_atoms = input_atoms
         self.preset = preset
         self.use_custodian = use_custodian
-        self.incar_copilot = incar_copilot
+        self.incar_copilot_mode = incar_copilot_mode
         self.copy_magmoms = copy_magmoms
         self.preset_mag_default = preset_mag_default
         self.mag_cutoff = mag_cutoff
@@ -315,12 +315,14 @@ class Vasp(Vasp_):
         self.user_calc_params = get_param_swaps(
             self.user_calc_params,
             self.input_atoms,
-            incar_copilot_mode=self.incar_copilot,
+            incar_copilot_mode=self.incar_copilot_mode,
             pmg_kpts=self.pmg_kpts,
         )
+        if self.incar_copilot_mode.lower() not in {"off", "critical"}:
+            self.user_calc_params = remove_unused_flags(self.user_calc_params)
 
         # Clean up the user calc parameters
-        self.user_calc_params = sort_dict(remove_unused_flags(self.user_calc_params))
+        self.user_calc_params = sort_dict(self.user_calc_params)
 
     def _run(
         self,