+++ src/quacc/recipes/vasp/mp_aloe.py
@@ -9,6 +9,8 @@ from quacc.calculators.vasp.params import MPtoASEConverter
 from quacc.recipes.vasp._base import run_and_summarize
 
 if TYPE_CHECKING:
+    from typing import Any
+
     from ase.atoms import Atoms
 
     from quacc.types import SourceDirectory, VaspSchema
@@ -16,7 +18,7 @@ if TYPE_CHECKING:
 
 @job
 def mp_aloe_static_job(
-    atoms: Atoms, prev_dir: SourceDirectory | None = None, **calc_kwargs
+    atoms: Atoms, prev_dir: SourceDirectory | None = None, **calc_kwargs: Any
 ) -> VaspSchema:
     """
     Run a static calculation with MP-ALOE settings.
@@ -45,7 +47,7 @@ def mp_aloe_static_job(
     calc_defaults = MPtoASEConverter(atoms=atoms, prev_dir=prev_dir).convert_input_set(
         MP24RelaxSet()
     )
-    calc_defaults |= {"kspacing": 0.2, "nsw": 0, "incar_copilot": "critical"}
+    calc_defaults |= {"kspacing": 0.2, "nsw": 0, "incar_copilot_mode": "critical"}
 
     return run_and_summarize(
         atoms,